pyproject.toml
[project]
name = "understanding-ai-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.4,<3.14"
dependencies = [
    "einops>=0.8.1",
    "ipykernel>=6.29.5",
    "kaleido>=1.0.0",
    "llama-index>=0.12.43",
    "llama-index-embeddings-huggingface>=0.5.5",
    "llama-index-readers-youtube-transcript>=0.3.0",
    "llama-index-vector-stores-qdrant>=0.6.0",
    "mcp>=1.10.1",
    "nbformat>=5.10.4",
    "plotly>=6.2.0",
    "python-dotenv (>=1.1.1,<2.0.0)",
]
